chore: remove wordpress api and env vars
