+++ .env.example
@@ -1,5 +1,4 @@
 # Example environment variables. Copy to .env and replace placeholders with real credentials.
 # Never commit real secrets to version control.
-VITE_WP_BASE_URL=https://renownedcomic.com/wp-admin/
-VITE_WP_USERNAME=your_username
-VITE_WP_APP_PASSWORD=your_app_password
+VITE_SUPABASE_URL=https://your-project.supabase.co
+VITE_SUPABASE_ANON_KEY=your_anon_key
